[lab-05] prepare lab

+++ include/lab_structures.hpp
@@ -3,13 +3,14 @@
 #ifndef INCLUDE_LAB_STRUCTURES_HPP_
 #define INCLUDE_LAB_STRUCTURES_HPP_
 
+#include <cstdint>
 #include <string>
 #include <vector>
 
 struct Student {
   std::string Name;
   std::string GroupId;
-  std::vector<unsigned> Ratings;
+  std::vector<uint8_t> Ratings;
   std::vector<std::string> Subjects;
 };
